refactor: move MockInputOutput class
